requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
outcome==1.2.0
psycopg2==2.9.8
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
starlette==0.27.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
wsproto==1.2.0
